backend/pyproject.toml
[project]
name = "stage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.3",
    "psycopg[binary]>=3.3.3",
    "pydantic[email]>=2.13.0",
    "pypdf>=6.10.1",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
